// user_microservice/src/protos/proto_gen_files/user.ts
export interface GetUserProfileRequest {
  userId: number;
  whoFind: number;
}
// ...
export interface GetUserProfileResponse {
  message: string;
  status: number;
}
// ...
export interface GetPasswordUserRequest {
  userId: number;
}

export interface GetPasswordUserResponse {
  hashedPassword: string;
}

export interface RemoveAccountRequest {
  userId: number;
}

export interface RemoveAccountResponse {
  message: string;
}

export interface RemoveArrayChatRequest {
  chatId: string;
  data: ArrayLinkUsers[];
}

export interface ArrayLinkUsers {
  userId: number;
}

export interface RemoveArrayChatResponse {
  message: string;
  status: number;
}

export interface AddChatToUserRequest {
  userId: number;
  chatId: string;
}

export interface AddChatToUserResponse {
  info: ResponseMessage | undefined;
}

export interface RemoveChatFromUserRequest {
  userId: number;
  chatId: string;
}

export interface RemoveChatFromUserResponse {
  info: ResponseMessage | undefined;
}

export interface CreateNewUserRequest {
  email: string;
  passwordHash: string;
  phoneNumber: string;
  username: string;
}

export interface CreateNewUserResponse {
  info: ResponseMessage | undefined;
}

export interface FindUserByIdRequest {
  userId: number;
}

export interface FindUserByIdResponse {
  userData?: UserDataId | undefined;
  notFound?: UserNotFoundResponse | undefined;
}

export interface UserDataId {
  userId: number;
  phoneNumber: string;
  email: string;
  passwordHash: string;
  username: string;
  chatReferences: string[];
}

export interface FindUserByUsernameRequest {
  username: string;
}

export interface FindUserByUsernameResponse {
  users: UserArray[];
}

export interface FindUserByEmailRequest {
  email: string;
}

export interface FindUserByEmailResponse {
  userData?: UserData | undefined;
  notFound?: UserNotFoundResponse | undefined;
}

export interface FindUserByPhoneNumberRequest {
  phoneNumber: string;
}

export interface FindUserByPhoneNumberResponse {
  userData?: UserData | undefined;
  notFound?: UserNotFoundResponse | undefined;
}
// ...
export const FindUserByIdRequest: MessageFns<FindUserByIdRequest> = {
  encode(message: FindUserByIdRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userId !== 0) {
      writer.uint32(8).int32(message.userId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FindUserByIdRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFindUserByIdRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.userId = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FindUserByIdRequest {
    return { userId: isSet(object.userId) ? globalThis.Number(object.userId) : 0 };
  },

  toJSON(message: FindUserByIdRequest): unknown {
    const obj: any = {};
    if (message.userId !== 0) {
      obj.userId = Math.round(message.userId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FindUserByIdRequest>, I>>(base?: I): FindUserByIdRequest {
    return FindUserByIdRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FindUserByIdRequest>, I>>(object: I): FindUserByIdRequest {
    const message = createBaseFindUserByIdRequest();
    message.userId = object.userId ?? 0;
    return message;
  },
};

function createBaseFindUserByIdResponse(): FindUserByIdResponse {
  return { userData: undefined, notFound: undefined };
}

export const FindUserByIdResponse: MessageFns<FindUserByIdResponse> = {
  encode(message: FindUserByIdResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userData !== undefined) {
      UserDataId.encode(message.userData, writer.uint32(10).fork()).join();
    }
    if (message.notFound !== undefined) {
      UserNotFoundResponse.encode(message.notFound, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FindUserByIdResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFindUserByIdResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.userData = UserDataId.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.notFound = UserNotFoundResponse.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FindUserByIdResponse {
    return {
      userData: isSet(object.userData) ? UserDataId.fromJSON(object.userData) : undefined,
      notFound: isSet(object.notFound) ? UserNotFoundResponse.fromJSON(object.notFound) : undefined,
    };
  },

  toJSON(message: FindUserByIdResponse): unknown {
    const obj: any = {};
    if (message.userData !== undefined) {
      obj.userData = UserDataId.toJSON(message.userData);
    }
    if (message.notFound !== undefined) {
      obj.notFound = UserNotFoundResponse.toJSON(message.notFound);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FindUserByIdResponse>, I>>(base?: I): FindUserByIdResponse {
    return FindUserByIdResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FindUserByIdResponse>, I>>(object: I): FindUserByIdResponse {
    const message = createBaseFindUserByIdResponse();
    message.userData = (object.userData !== undefined && object.userData !== null)
      ? UserDataId.fromPartial(object.userData)
      : undefined;
    message.notFound = (object.notFound !== undefined && object.notFound !== null)
      ? UserNotFoundResponse.fromPartial(object.notFound)
      : undefined;
    return message;
  },
};
// ...
export const FindUserByEmailRequest: MessageFns<FindUserByEmailRequest> = {
  encode(message: FindUserByEmailRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.email !== "") {
      writer.uint32(10).string(message.email);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FindUserByEmailRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFindUserByEmailRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.email = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FindUserByEmailRequest {
    return { email: isSet(object.email) ? globalThis.String(object.email) : "" };
  },

  toJSON(message: FindUserByEmailRequest): unknown {
    const obj: any = {};
    if (message.email !== "") {
      obj.email = message.email;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FindUserByEmailRequest>, I>>(base?: I): FindUserByEmailRequest {
    return FindUserByEmailRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FindUserByEmailRequest>, I>>(object: I): FindUserByEmailRequest {
    const message = createBaseFindUserByEmailRequest();
    message.email = object.email ?? "";
    return message;
  },
};

function createBaseFindUserByEmailResponse(): FindUserByEmailResponse {
  return { userData: undefined, notFound: undefined };
}

export const FindUserByEmailResponse: MessageFns<FindUserByEmailResponse> = {
  encode(message: FindUserByEmailResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userData !== undefined) {
      UserData.encode(message.userData, writer.uint32(10).fork()).join();
    }
    if (message.notFound !== undefined) {
      UserNotFoundResponse.encode(message.notFound, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FindUserByEmailResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFindUserByEmailResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.userData = UserData.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.notFound = UserNotFoundResponse.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FindUserByEmailResponse {
    return {
      userData: isSet(object.userData) ? UserData.fromJSON(object.userData) : undefined,
      notFound: isSet(object.notFound) ? UserNotFoundResponse.fromJSON(object.notFound) : undefined,
    };
  },

  toJSON(message: FindUserByEmailResponse): unknown {
    const obj: any = {};
    if (message.userData !== undefined) {
      obj.userData = UserData.toJSON(message.userData);
    }
    if (message.notFound !== undefined) {
      obj.notFound = UserNotFoundResponse.toJSON(message.notFound);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FindUserByEmailResponse>, I>>(base?: I): FindUserByEmailResponse {
    return FindUserByEmailResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FindUserByEmailResponse>, I>>(object: I): FindUserByEmailResponse {
    const message = createBaseFindUserByEmailResponse();
    message.userData = (object.userData !== undefined && object.userData !== null)
      ? UserData.fromPartial(object.userData)
      : undefined;
    message.notFound = (object.notFound !== undefined && object.notFound !== null)
      ? UserNotFoundResponse.fromPartial(object.notFound)
      : undefined;
    return message;
  },
};
// ...
export interface UserService {
  CreateNewUser(request: CreateNewUserRequest): Promise<CreateNewUserResponse>;
  FindUserById(request: FindUserByIdRequest): Promise<FindUserByIdResponse>;
  FindUserByUsername(request: FindUserByUsernameRequest): Promise<FindUserByUsernameResponse>;
  FindUserByEmail(request: FindUserByEmailRequest): Promise<FindUserByEmailResponse>;
  FindUserByPhoneNumber(request: FindUserByPhoneNumberRequest): Promise<FindUserByPhoneNumberResponse>;
  AddChatToUser(request: AddChatToUserRequest): Promise<AddChatToUserResponse>;
  RemoveChatFromUser(request: RemoveChatFromUserRequest): Promise<RemoveChatFromUserResponse>;
  RemoveArrayChat(request: RemoveArrayChatRequest): Promise<RemoveArrayChatResponse>;
  RemoveAccount(request: RemoveAccountRequest): Promise<RemoveAccountResponse>;
  GetPasswordUser(request: GetPasswordUserRequest): Promise<GetPasswordUserResponse>;
  GetUserProfile(request: GetUserProfileRequest): Promise<GetUserProfileResponse>;
  UpdateUserProfile(request: UpdateUserProfileRequest): Promise<UpdateUserProfileResponse>;
  UpdateUserPassword(request: UpdateUserPasswordRequest): Promise<UpdateUserPasswordResponse>;
  ToggleUserProfileCheck(request: ToggleUserProfileCheckRequset): Promise<ToggleUserProfileCheckResponse>;
  DeleteAvatarUser(request: DeleteAvatarUserRequest): Promise<DeleteAvatarUserResponse>;
  FindUserAvatar(request: FindUserAvatarRequest): Promise<FindUserAvatarResponse>;
  FindUserAvatarArray(request: FindUserAvatarArrayRequest): Promise<FindUserAvatarArrayResponse>;
  UploadAvatarUser(request: UploadAvatarUserRequest): Promise<UploadAvatarUserResponse>;
}